+++ tests/TestEnvironmentManager.php
@@ -87,7 +87,7 @@ class TestEnvironmentManager
             $log->pushProcessor(new IntrospectionProcessor);
             $log->pushProcessor(new UidProcessor());
 
-            $log->pushHandler(new \Monolog\Handler\StreamHandler(self::LOG_FILE, Logger::DEBUG));
+            $log->pushHandler(new \Monolog\Handler\StreamHandler(__DIR__ . '/logs/' . self::LOG_FILE, Logger::DEBUG));
             $log->pushHandler(new \Monolog\Handler\StreamHandler('php://stdout', Logger::WARNING));
         }