+++ examples/user.add.php
@@ -2,21 +2,13 @@
 declare(strict_types=1);
 require_once __DIR__ . '/init.php';
 
-$userArray = [
-    'name' => 'Михаил',
-    'login' => 'grishi+1@rarus.ru',
-    'phone' => '+7978 888 22 22',
-    'email' => 'grishi@rarus.ru',
-    'gender' => 'male',
-    'birthdate' => '742262400',
-];
-
-
 $newUser = \Rarus\BonusServer\Users\DTO\Fabric::createNewInstance(
     'grishi-' . random_int(0, PHP_INT_MAX),
     'Михаил Гришин',
     '+7978 888 22 22',
-    'grishi@rarus.ru'
+    'grishi@rarus.ru',
+    null,
+    new \DateTime('06.06.1985')
 );
 
 $transport = \Rarus\BonusServer\Users\Transport\Role\Organization\Fabric::getInstance($apiClient, new \Money\Currency('RUB'), $log);