+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # bonus-server-php-sdk change log
 
+## 0.5.2 (27.08.2018)
+- исправлена ошибка c часовыми поясами и добавлены проверки в тесты `Rarus\BonusServer\Util\DateTimeParser` [issue#58](https://github.com/rarus/bonus-server-php-sdk/issues/58)
+
 ## 0.5.1 (20.08.2018)
 - исправлена ошибка с точностью в парсере timestamp `Rarus\BonusServer\Util\DateTimeParser` [issue#55](https://github.com/rarus/bonus-server-php-sdk/issues/55)
 - исправлена ошибка с некорректными датами рождения при добавлении новых пользователей [issue#55](https://github.com/rarus/bonus-server-php-sdk/issues/55)